refactor : 요청 request Dto에 memberId만 사용하는 중복 클래스들 제거 후 클래스 하나로 통일
