--- src/main/java/dnd/myOcean/domain/letter/dto/request/LetterPassRequest.java
@@ -1,16 +1,0 @@
-package dnd.myOcean.domain.letter.dto.request;
-
-import jakarta.validation.constraints.Null;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LetterPassRequest {
-
-    @Null
-    private Long memberId;
-}
