--- src/main/java/dnd/myOcean/domain/letter/dto/request/LetterReadRequest.java
@@ -1,17 +1,0 @@
-package dnd.myOcean.domain.letter.dto.request;
-
-
-import jakarta.validation.constraints.Null;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LetterReadRequest {
-
-    @Null
-    private Long memberId;
-}
