--- src/main/java/dnd/myOcean/domain/member/dto/request/WorryDeleteRequest.java
@@ -1,16 +1,0 @@
-package dnd.myOcean.domain.member.dto.request;
-
-import jakarta.validation.constraints.Null;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class WorryDeleteRequest {
-
-    @Null
-    private Long memberId;
-}
