+++ src/main/java/dnd/myOcean/global/auth/aop/dto/CurrentMemberIdRequest.java
@@ -1,4 +1,5 @@
-package dnd.myOcean.domain.letter.dto.request;
+package dnd.myOcean.global.auth.aop.dto;
+
 
 import jakarta.validation.constraints.Null;
 import lombok.AccessLevel;
@@ -9,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LetterDeleteRequest {
+public class CurrentMemberIdRequest {
 
     @Null
     private Long memberId;
